refactor: rename scripts, examples and tests to .js
