+++ .scripts/move-artifact.js
