+++ tests/WaveShaper.spec.js
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import { AudioContext } from '../index.mjs';
+import { AudioContext } from '../index.js';
 
 describe('# WaveShaper', () => {
   describe('## curve', () => {
