+++ tests/cast.spec.js
