+++ tests/test-offline-context-gc.js
@@ -1,11 +1,7 @@
-import { OfflineAudioContext } from '../index.mjs';
+import { OfflineAudioContext } from '../index.js';
 
 // Regression test for #133
 // <https://github.com/ircam-ismm/node-web-audio-api/issues/133>
-//
-// uncomment Drop trait in `offline_audio_context.rs` and rebuild
-// `node --trace-gc --expose-gc tests/junk-test-offline-context-gc.mjs`
-// @todo - add a build flag to automate this
 
 for( let i=0; i < 100000; i++ ) {
   console.log('+ i:', i);
