+++ tests/worklets/invalid-ctor.worklet.js
