--- tests/worklets/invalid.worklet.mjs
@@ -1,1 +1,0 @@
-const a = invalid;
